Vendor require_python().
