+++ setup.py
@@ -1,11 +1,39 @@
 #!/usr/bin/env python3
 
 from debian.changelog import Changelog
-from setup_helpers import require_python
 from setuptools import setup
 
 
+def require_python(minimum):
+    """Require at least a minimum Python version.
+
+    The version number is expressed in terms of `sys.hexversion`.  E.g. to
+    require a minimum of Python 2.6, use::
+
+    >>> require_python(0x206000f0)
+
+    :param minimum: Minimum Python version supported.
+    :type minimum: integer
+    """
+    if sys.hexversion < minimum:
+        hversion = hex(minimum)[2:]
+        if len(hversion) % 2 != 0:
+            hversion = '0' + hversion
+        split = list(hversion)
+        parts = []
+        while split:
+            parts.append(int(''.join((split.pop(0), split.pop(0))), 16))
+        major, minor, micro, release = parts
+        if release == 0xf0:
+            print('Python {}.{}.{} or better is required'.format(
+                major, minor, micro))
+        else:
+            print('Python {}.{}.{} ({}) or better is required'.format(
+                major, minor, micro, hex(release)[2:]))
+        sys.exit(1)
+
 require_python(0x30500f0)
+
 
 with open('debian/changelog', encoding='utf-8') as fp:
     __version__ = str(Changelog(fp).get_version())
